Refactors classes into separate files.
